main: Move ready to own file
