+++ lxd/main.go
@@ -244,30 +244,6 @@ func run() error {
 	return cmdDaemon()
 }
 
-func cmdReady() error {
-	c, err := lxd.NewClient(&lxd.DefaultConfig, "local")
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequest("PUT", c.BaseURL+"/internal/ready", nil)
-	if err != nil {
-		return err
-	}
-
-	raw, err := c.Http.Do(req)
-	if err != nil {
-		return err
-	}
-
-	_, err = lxd.HoistResponse(raw, lxd.Sync)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
 func cmdShutdown() error {
 	var timeout int
